Word2Vec Approach:
Previous approach trained word embeddings along with model parameters. This resulted in very large gradients and difficulty finding the local minimum.
Word2Vec is a powerful approach for building machine understanding of a language space. The current approach is to train an embedding in word2vec and feed data interpreted through this embedding into the neural network. This should decrease the overhead loss and let the network focus on interpreting sequence relations while having semantic relations already done in preprocessing.